lily: fix some type conversion warnings

+++ lily/lookup.cc
@@ -311,7 +311,7 @@ Lookup::round_filled_polygon (vector<Offset> const &points,
 
       for (vsize i = 0; i < points.size (); i++)
         {
-          int i0 = i;
+          int i0 = (int) i;
           int i1 = (i + 1) % points.size ();
           int i2 = (i + 2) % points.size ();
           Offset p0 = points[i0];
